Create chats.py
Process a text and distill statistics about the bi-gram and tri-gram word occurrences in the entered text. It does this by using a dictionary of words and bi-grams. Each dictionary entry gives the list of words in the text that follow that word/bi-gram (and possibly a count of the number of times each follower does so). The code then generates random text based on these statistics.